Refactor component composition to use ThemeProvider on pages directly
