+++ src/views/Homepage/Homepage.tsx
@@ -3,13 +3,13 @@ import Typography from "@material-ui/core/Typography";
 import Button from "@material-ui/core/Button";
 import { Link as RouterLink } from "react-router-dom";
 
-import Navbar from "../components/Navbar";
-import ContentBlock from "../components/ContentBlock";
-import FeaturedOrgList from "../components/FeaturedOrgList";
-import OrgList from "../components/OrgList/OrgList";
-import Subscribe from "../components/Subscribe";
+import Navbar from "../../components/Navbar";
+import ContentBlock from "../../components/ContentBlock";
+import FeaturedOrgList from "../../components/FeaturedOrgList";
+import OrgList from "../../components/OrgList/OrgList";
+import Subscribe from "../../components/Subscribe";
 
-import mockData from "../mockData";
+import mockData from "../../mockData";
 
 function Homepage() {
   return (
