+++ src/views/OrgPage/index.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+import { ThemeProvider } from "@material-ui/core/styles";
+
+import theme from "../../theme";
+import OrgPage from "./OrgPage";
+
+export default () => (
+  <ThemeProvider theme={theme}>
+    <OrgPage />
+  </ThemeProvider>
+);
